Moves getRandomPokemon to separate file
