+++ src/Components/Header/Header.scss
@@ -1,5 +1,5 @@
 header {
-  background: #FF0000;
+  background: linear-gradient(to left, #FF0000 4%, #CC0000 100%);
   height: 23vh;
   display: flex;
   flex-flow: column;
